working increment/set, not hooked up yet

+++ sqlite/src/comdb2build.h
@@ -117,7 +117,7 @@ void comdb2grant(Parse* pParse, int revoke, int permission, Token* nm,
 
 void comdb2timepartRetention(Parse*, Token*, Token*, int val);
 void comdb2CounterIncr(Parse*, Token*, Token*);
-void comdb2CounterSet(Parse*, Token*, Token*, int val);
+void comdb2CounterSet(Parse*, Token*, Token*, long long val);
 
 void comdb2enableAuth(Parse* pParse, int on);
 void comdb2setPassword(Parse* pParse, Token* password, Token* nm);
